Split the SDL graphics contexts into their own files.
